requirements.txt
langchain
openai
pytest
fastapi
uvicorn
python-dotenv

